--- a/src/DynamicExpresso.Core/Reflection/MemberFinder.cs
+++ b/src/DynamicExpresso.Core/Reflection/MemberFinder.cs
@@ -9,15 +9,61 @@ namespace DynamicExpresso.Reflection
 {
 	internal class MemberFinder
 	{
+		private struct MemberTypeKey
+		{
+			public Type _type;
+			public string _memberName;
+			public BindingFlags _flags;
+
+			public MemberTypeKey(Type type, string memberName, BindingFlags flags)
+			{
+				_type = type;
+				_memberName = memberName;
+				_flags = flags;
+			}
+		}
+
+		private class MemberTypeKeyEqualityComparer : IEqualityComparer<MemberTypeKey>
+		{
+			public static readonly MemberTypeKeyEqualityComparer Instance = new MemberTypeKeyEqualityComparer();
+
+			public bool Equals(MemberTypeKey x, MemberTypeKey y)
+			{
+				return x._type.Equals(y._type) && x._flags == y._flags && x._memberName.Equals(y._memberName, StringComparison.Ordinal);
+			}
+
+
+			public int GetHashCode(MemberTypeKey obj)
+			{
+				return obj._type.GetHashCode() ^ obj._flags.GetHashCode() ^ obj._memberName.GetHashCode();
+			}
+		}
+
+		private class MemberTypeKeyCaseInsensitiveEqualityComparer : IEqualityComparer<MemberTypeKey>
+		{
+			public static readonly MemberTypeKeyCaseInsensitiveEqualityComparer Instance = new MemberTypeKeyCaseInsensitiveEqualityComparer();
+			public bool Equals(MemberTypeKey x, MemberTypeKey y)
+			{
+				return x._type.Equals(y._type) && x._flags == y._flags && x._memberName.Equals(y._memberName, StringComparison.OrdinalIgnoreCase);
+			}
+
+			public int GetHashCode(MemberTypeKey obj)
+			{
+				return obj._type.GetHashCode() ^ obj._flags.GetHashCode() ^ StringComparer.InvariantCulture.GetHashCode(obj._memberName);
+			}
+		}
+
 		private readonly ParserArguments _arguments;
 		private readonly BindingFlags _bindingCase;
 		private readonly MemberFilter _memberFilterCase;
+		private readonly Dictionary<MemberTypeKey, MemberInfo[]> _members;
 
 		public MemberFinder(ParserArguments arguments)
 		{
 			_arguments = arguments;
 			_bindingCase = arguments.Settings.CaseInsensitive ? BindingFlags.IgnoreCase : BindingFlags.Default;
 			_memberFilterCase = arguments.Settings.CaseInsensitive ? Type.FilterNameIgnoreCase : Type.FilterName;
+			_members = new Dictionary<MemberTypeKey, MemberInfo[]>(arguments.Settings.CaseInsensitive ? (IEqualityComparer<MemberTypeKey>)MemberTypeKeyCaseInsensitiveEqualityComparer.Instance : MemberTypeKeyEqualityComparer.Instance);
 		}
 
 		public MemberInfo FindPropertyOrField(Type type, string memberName, bool staticAccess)
@@ -25,22 +71,61 @@ namespace DynamicExpresso.Reflection
 			var flags = BindingFlags.Public | BindingFlags.DeclaredOnly |
 						(staticAccess ? BindingFlags.Static : BindingFlags.Instance) | _bindingCase;
 
-			foreach (var t in SelfAndBaseTypes(type))
+			//Not looping here because FindMembers loops internally and returns us the highest level member
+			var members = FindMembers(type, MemberTypes.Property | MemberTypes.Field, flags, memberName);
+			if (members.Length != 0)
 			{
-				var members = t.FindMembers(MemberTypes.Property | MemberTypes.Field, flags, _memberFilterCase, memberName);
-				if (members.Length != 0)
-					return members[0];
+				return members[0];
 			}
 			return null;
+		}
+
+		private MemberInfo[] FindMembers(Type type, MemberTypes memberTypes, BindingFlags flags, string memberName)
+		{
+			var key = new MemberTypeKey(type, memberName, flags);
+			if (_members.TryGetValue(key, out var members))
+			{
+				if (members != null)
+				{
+					return members;
+				}
+			}
+			members = type.FindMembers(memberTypes, flags, _memberFilterCase, memberName);
+			if (members.Length == 0)
+			{
+				foreach (var v in SelfAndBaseTypes(type))
+				{
+					if (v == type)
+					{
+						continue;
+					}
+					var subMembers = FindMembers(v, memberTypes, flags, memberName);
+					if (members.Length == 0 && subMembers.Length > 0)
+					{
+						//We don't break here because there is a possibility that somebody outside here is also doing an additional tree prioritization (See FindMethods)
+						members = subMembers;
+					}
+				}
+			}
+			if (members.Length == 0)
+			{
+				members = Array.Empty<MemberInfo>();
+			}
+			_members[key] = members;
+			return members;
 		}
 
 		public IList<MethodData> FindMethods(Type type, string methodName, bool staticAccess, Expression[] args)
 		{
 			var flags = BindingFlags.Public | BindingFlags.DeclaredOnly |
 						(staticAccess ? BindingFlags.Static : BindingFlags.Instance) | _bindingCase;
+			//Yes, FindMembers loops internally, but there are certain circumstances where we end up also needing this
+			//An example is Type.GetMethods(), where there are multiple overloads on the base Type class
+			//and one override on a derived class.  The DeclaredOnly flag will return the single method on the derived class unless we go lower.
+			//The existing unit tests do pass if we get rid of DeclaredOnly and remove the loop, but I'm not sure of the actual impact that may have.
 			foreach (var t in SelfAndBaseTypes(type))
 			{
-				var members = t.FindMembers(MemberTypes.Method, flags, _memberFilterCase, methodName);
+				var members = FindMembers(t, MemberTypes.Method, flags, methodName);
 				var applicableMethods = MethodResolution.FindBestMethod(members.Cast<MethodBase>(), args);
 
 				if (applicableMethods.Count > 0)
@@ -100,14 +185,19 @@ namespace DynamicExpresso.Reflection
 		{
 			if (type.IsInterface)
 			{
-				var types = new List<Type>();
-				AddInterface(types, type);
+				var types = new HashSet<Type>();
+				foreach (var v in AddInterface(types, type))
+				{
+					yield return v;
+				}
 
-				types.Add(typeof(object));
-
-				return types;
+				yield return typeof(object);
+				yield break;
 			}
-			return SelfAndBaseClasses(type);
+			foreach (var t in SelfAndBaseClasses(type))
+			{
+				yield return t;
+			}
 		}
 
 		private static IEnumerable<Type> SelfAndBaseClasses(Type type)
@@ -119,14 +209,14 @@ namespace DynamicExpresso.Reflection
 			}
 		}
 
-		private static void AddInterface(List<Type> types, Type type)
+		private static IEnumerable<Type> AddInterface(HashSet<Type> types, Type type)
 		{
 			if (!types.Contains(type))
 			{
-				types.Add(type);
-				foreach (var t in type.GetInterfaces())
+				yield return type;
+				foreach (var i in type.GetInterfaces().SelectMany(x => AddInterface(types, x)))
 				{
-					AddInterface(types, t);
+					yield return i;
 				}
 			}
 		}