As promised, in #349 I finally have my contribution to the performance wrapped up and ready to go.

The proposed changes break down into addressing the following profiling findings

1. A lot of time was spent in the call to ErrorMessages.[SomeErrorMessage] that was being provided to generic helper methods
2. HasParamsArrayType is in the hot path and spent a lot of time accessing code invisible to us
3. MemberFinder using the Type.FindMembers method also spent a good proportion of the time accessing code invisible to us

I also had the goal of improving the FindBestMethods logic to reduce List creation as much as possible.  Due to a couple of tests hitting very specific asserts I was unable to get rid of the List completely but the single list that is allocated will only have more than 2 items in extremely rare cases, whereas the old logic regularly could hit 10 or 11 that it then has to reprocess at the end.  Additionally, if we have already hit an EXACT match (no type converting, no params usage, etc.) of a method and a candidate method requires some kind of conversion we end the preparing of that method altogether.

Going forward I would like to break the usage of CheckMethodMatchAndPrepareIt logic into the two separate stages
1. Gather the best possible matching methods without any of the extra hinting that preparing adds to the process.
2. Prepare the methods and prune if needed.
However at this time I didn't have the time to really dig into that logic and figure out exactly when/why things happen.

The only thing I'm not sure on is if we would like the MemberFinder's caching to be something people can turn on/off.  I was thinking it is fine to keep on and not even expose that it is happening since the caching only lives per Eval or Parse call and does not persist across multiple calls (unlike the Ncalc library that statically caches its built expressions)

  My timings from the same benchmarking logic as show in #349 after improvements is

<!--StartFragment-->
Method | N | Mean | Error | StdDev
-- | -- | -- | -- | --
DynamicExpressoParsingOnly | 5 | 10,246.8 ns | 200.6 ns | 281.22 ns
DynamicExpressoEval | 5 | 184830.1 ns | 3602.97 ns | 5167.27 ns
NcalcEval | 5 | 784.3 ns | 15.46 ns | 26.25 ns
NcalcEvalNoCache | 5 | 11,609.5 ns | 202.24 ns | 179.28 ns
DynamicExpressoParsingOnly | 20 | 63,925.0 ns | 286.82 ns | 239.50 ns
DynamicExpressoEval | 20 | 314,246.6 ns | 1,869.79 ns | 1,561.36 ns
NcalcEval | 20 | 2,798.6 ns | 12.05 ns | 10.06 ns
NcalcEvalNoCache | 20 | 44130.1 ns | 398.75 ns | 372.99 ns

<!--EndFragment-->

<!--StartFragment-->
Timings from before my changes
Method | N | Mean | Error | StdDev     
-- | -- | -- | -- | --
| DynamicExpressoParsingOnly | 5  |  13,778.0 ns |   223.15 ns |   197.81 ns
| DynamicExpressoEval        | 5  | 191,236.9 ns | 3,764.84 ns | 5,635.04 ns
| NcalcEval                  | 5  |     755.8 ns |    14.41 ns |    15.42 ns
| NcalcEvalNoCache           | 5  |  11,846.7 ns |   232.50 ns |   258.42 ns
| DynamicExpressoParsingOnly | 20 |  57,399.6 ns |   576.39 ns |   481.31 ns
| DynamicExpressoEval        | 20 | 334,871.9 ns | 5,342.14 ns | 4,997.04 ns
| NcalcEval                  | 20 |   3,027.6 ns |    57.66 ns |    61.70 ns
| NcalcEvalNoCache           | 20 |  43,733.4 ns |   585.24 ns |   518.80 ns

<!--EndFragment-->


I'm not sure what's up with the DynamicExpressoParsingOnly N = 20 case since the Eval is proportionally faster.

Once this PR is merged in I can open up a new one with my changes.